David Baxters version of the key pruning script
This is the laternative form of the script, modified by David baxter in 2016 to create keys for Keybase from Eflora_treatment.txt.  I am not sure what the differences are, but all the instructions point to using Dick's script.  David used this version for the last upload of the keybase keys